Reworked device discovery and init
* try to figue out device dependencies from the device tree

+++ kernel/kernel/proc.c
@@ -833,7 +833,7 @@ void debug_print_process_list(bool print_call_stack_user,
         [UNUSED] "unused",     [USED] "used",       [SLEEPING] "sleeping",
         [RUNNABLE] "runnable", [RUNNING] "running", [ZOMBIE] "zombie"};
 
-    printk("\nProcess list (%zd)\n", smp_processor_id());
+    printk("\nProcess list (%zd)\n", (size_t)smp_processor_id());
     for (struct process *proc = g_process_list;
          proc < &g_process_list[MAX_PROCS]; proc++)
     {
